move methods to a folder
