+++ src/libinjector/Makefile.am
@@ -102,10 +102,14 @@
 #                                                                         #
 #*************************************************************************#
 
-h_sources = libinjector.h private.h win_private.h linux/linux_utils.h linux/linux_debug.h linux/linux_shellcode.h linux/linux_syscalls.h
-c_sources = injector.c win_injector.c linux/linux_injector.c linux/linux_utils.c linux/linux_debug.c linux/linux_shellcode.c linux/linux_syscalls.c injector_stack.c
+h_sources = libinjector.h private.h win_private.h \
+	linux/linux_utils.h linux/linux_debug.h linux/linux_syscalls.h \
+	linux/methods/linux_shellcode.h
+c_sources = injector.c win_injector.c injector_stack.c \
+	linux/linux_injector.c linux/linux_utils.c linux/linux_debug.c linux/linux_syscalls.c \
+	linux/methods/linux_shellcode.c
 
-AM_CPPFLAGS = -I$(top_srcdir) -I$(top_srcdir)/src -I$(srcdir)
+AM_CPPFLAGS = -I$(top_srcdir) -I$(top_srcdir)/src -I$(srcdir) -I$(srcdir)/linux
 AM_CPPFLAGS += $(VMI_CFLAGS)
 AM_CPPFLAGS += $(GLIB_CFLAGS)
