+++ src/libinjector/linux/methods/linux_shellcode.h
